adjusted to use long double, so that rendering doesn't plateau as quickly.  Also adjusted the center finding, little jerky, but avoids getting trapped in a bad area

+++ c/fractal.h
@@ -1,4 +1,4 @@
 #ifndef FRACTAL_H
 #define FRACTAL_H
-double *mandelbrot(float *xs, float *ys, int height, int width, int maxIter);
+double *mandelbrot(long double *xs, long double *ys, int height, int width, int maxIter);
 #endif // FRACTAL_H
